Refactored old IOTATest. It's now the ClientTest which aligns to the recent renaming activities in the lib.
